Fix linter problems
