+++ example/i2c_gesture_sensor.dart
@@ -10,7 +10,7 @@ import 'package:dart_periphery/dart_periphery.dart';
 ///
 void main() {
   // Select the right I2C bus number /dev/i2c-?
-  // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian)
+  // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian), 4 BPI-F3
   var i2c = I2C(1);
 
   try {
